server/requirements.txt
Flask==3.0.2
Flask-APScheduler==1.13.1
Flask-Cors>=4.0.2
Flask-Limiter==3.9.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Talisman==1.1.0
Pillow==11.0.0
python-dotenv==1.0.1
requests>=2.32.0
requests-oauthlib==1.3.1
selenium==4.25.0
Werkzeug==3.1.3
gunicorn==23.0.0
boto3==1.34.69